+++ lab5/app_url.txt
@@ -1,0 +1,1 @@
+https://go-lab5.herokuapp.com/
