+++ lab5/kiennam99/go.mod
@@ -1,0 +1,3 @@
+module Lab5
+
+go 1.17
